requirements.txt
telethon
python-dotenv
requests
flask
